Finally finished initial long 2016day11 solution
Started this problem back in 2017, left it alone for years, and then
came back to it today and cringed at my earlier code. A lot of the
code before the find_moves function is close to my original logic,
just with several optimizations I'm now aware of. The find_moves
function is entirely new, as my old process was a depth-first recursive
monstrosity.

This is still not as fast as I'd like, it took 12 seconds for part 1,
but 13-14 minutes for part 2 (also 5.4GB of RAM)