Generalized problem finding solution 2024day24
The following line is a hack as it identified a wrong wire first
first.startswith("z") and len(set(s_wire.get_inputs(ignore_wires))) > 9
We know that if we ignore all wires from the z bit above us then the
correct amount of wires involved with the solution for the current level
should be 9 unique wires. This was found through testing.